backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
lightgbm==4.3.0
websockets==12.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
joblib==1.3.2
